Fix clang tidy readability braces

+++ Source/utils/text_value_converter.h
@@ -25,7 +25,7 @@ public:
         return value > 0.5f ? translate("active") : translate("bypassed");
     }
 
-    float operator()(const String& text) const { return text == translate("active"); }
+    float operator()(const String& text) const { return static_cast<float>(text == translate("active")); }
 };
 
 class FrequencyTextConverter
